services/ai-agent/pyproject.toml
[project]
name = "nba-ai-agent"
version = "1.0.0"
description = "NBA Prediction Agent built with LangChain v1.x and LangGraph v1.x"
dependencies = [
    "langchain>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-openai>=1.0.0",
    "supabase>=2.0.0",
    "langchain-tavily>=0.2.17",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]


